Move functionality of Rack::Session::Redis to the Redis::Rack module
This allows all functionality of `Rack::Session::Redis` to be mixed into
other classes without needing to inherit from `Rack::Session::Redis`
itself. Pre-requisite for the fix in redis-store/redis-actionpack#22

Resolves #38
