+++ lib/redis-rack.rb
@@ -1,3 +1,1 @@
-require 'redis-store'
-require 'redis/rack/version'
-require 'rack/session/redis'
+require 'redis/rack'
